Cleanup AI Core code
