+++ lib/handlers/utils.js
@@ -1,3 +1,6 @@
+import cds from '@sap/cds';
+const LOG = cds.log('@cap-js/ai');
+
 export function getProperty(where, property) {
 	let res = null;
 	if (!where) return null;
@@ -19,4 +22,30 @@ export function getProperty(where, property) {
 		}
 	}
 	return null;
+}
+
+export async function parseResponse(req, response) {
+	if (response.ok) {
+		let res = await response.json();
+		if (res.resources) {
+			res.resources['$odata.count'] = res.count;
+			res = res.resources;
+		}
+		if (req.query.SELECT?.one) {
+			res = Array.isArray(res) ? res[0] : res;
+		}
+		return res;
+	} else {
+		const body = response.headers.get('content-type')?.match('json')
+			? JSON.stringify(await response.json())
+			: response.status;
+		LOG.error(
+			`Error when requesting ${req.target.name} from AI Core for tenant: `,
+			cds.context.tenant,
+			req.event,
+			req.query,
+			body
+		);
+		return {};
+	}
 }
